pyproject.toml
[project]
name = "agent-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb",
    "langchain-ollama",
    "langchain",
    "langchain-community",
    "ollama",
    "pypdf",
    "pytest",
    "python-dotenv",
    "streamlit",
    "tiktoken",
    "unstructured",
]
